+++ functions/run/run_func.go
@@ -13,6 +13,7 @@ import (
 	"github.com/vmware-tanzu/crash-diagnostics/functions/providers"
 	"github.com/vmware-tanzu/crash-diagnostics/functions/scriptconf"
 	"github.com/vmware-tanzu/crash-diagnostics/functions/sshconf"
+	"github.com/vmware-tanzu/crash-diagnostics/ssh"
 	"github.com/vmware-tanzu/crash-diagnostics/typekit"
 	"go.starlark.net/starlark"
 )
@@ -73,8 +74,51 @@ func runFunc(thread *starlark.Thread, _ *starlark.Builtin, _ starlark.Tuple, kwa
 		}
 	}
 
-	result := newCmd().Run(thread, agent, args)
+	result := Run(thread, agent, args)
 
 	// convert and return result
 	return functions.Result(Name, result)
+}
+
+// Run runs the command function
+func Run(t *starlark.Thread, agent ssh.Agent, args Args) Result {
+	if reflect.ValueOf(args.SSHConfig).IsZero() {
+		return Result{Error: "missing SSH config"}
+	}
+	sshConf := args.SSHConfig
+
+	if reflect.ValueOf(args.Resources).IsZero() {
+		return Result{Error: "missing resources"}
+	}
+	hosts := args.Resources.Hosts
+	if len(hosts) == 0 {
+		return Result{Error: fmt.Sprintf("%s provided no host", args.Resources.Provider)}
+	}
+
+	var cmdResults []RemoteProc
+	for _, host := range hosts {
+		var jumpProxy *ssh.ProxyJumpArgs
+		if sshConf.JumpHost != "" && sshConf.JumpUsername != "" {
+			jumpProxy = &ssh.ProxyJumpArgs{
+				User: sshConf.JumpUsername,
+				Host: sshConf.JumpHost,
+			}
+		}
+		sshArgs := ssh.SSHArgs{
+			User:           sshConf.Username,
+			Host:           host,
+			Port:           sshConf.Port,
+			MaxRetries:     int(sshConf.MaxRetries),
+			ProxyJump:      jumpProxy,
+			PrivateKeyPath: sshConf.PrivateKeyPath,
+		}
+
+		var errMsg string
+		sshOutput, err := ssh.Run(sshArgs, agent, args.Cmd)
+		if err != nil {
+			errMsg = err.Error()
+		}
+		cmdResults = append(cmdResults, RemoteProc{Error: errMsg, Host: host, Output: sshOutput})
+	}
+	return Result{Procs: cmdResults}
 }
