+++ functions/run_local/runlocal_func.go
@@ -6,6 +6,7 @@ package run_local
 import (
 	"fmt"
 
+	"github.com/vladimirvivien/echo"
 	"github.com/vmware-tanzu/crash-diagnostics/functions"
 	"github.com/vmware-tanzu/crash-diagnostics/functions/builtins"
 	"github.com/vmware-tanzu/crash-diagnostics/typekit"
@@ -31,8 +32,23 @@ func runLocalFunc(thread *starlark.Thread, b *starlark.Builtin, _ starlark.Tuple
 		return functions.Error(Name, fmt.Errorf("%s: %s", Name, err))
 	}
 
-	result := newCmd().Run(thread, args)
+	result := Run(thread, args)
 
 	// convert and return result
 	return functions.Result(Name, result)
+}
+
+// Run executes the command function
+func Run(t *starlark.Thread, args Args) Result {
+	proc := echo.New().RunProc(args.Cmd)
+	if proc.Err() != nil {
+		return Result{Error: proc.Err().Error()}
+	}
+	return Result{
+		Proc: LocalProc{
+			Pid:      int64(proc.ID()),
+			Result:   proc.Result(),
+			ExitCode: int64(proc.ExitCode()),
+		},
+	}
 }
