+++ functions/run_local/runlocal_run_test.go
@@ -25,18 +25,18 @@ func TestCmd_Run(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			result := newCmd().Run(&starlark.Thread{}, test.args)
+			result := Run(&starlark.Thread{}, test.args)
 			if !test.shouldFail && result.Error != "" {
 				t.Fatal(result.Error)
 			}
 
-			if result.Result != test.expected {
-				t.Errorf("command returned unexpected result: %s", result.Result)
+			if result.Proc.Result != test.expected {
+				t.Errorf("command returned unexpected result: %s", result.Proc.Result)
 			}
-			if result.Pid == 0 {
+			if result.Proc.Pid == 0 {
 				t.Errorf("successful command returned 0 pid")
 			}
-			if result.ExitCode != 0 {
+			if result.Proc.ExitCode != 0 {
 				t.Errorf("successful command returned non-zero exit code")
 			}
 		})
