+++ functions/run_local/types.go
@@ -7,9 +7,13 @@ type Args struct {
 	Cmd string `name:"cmd"`
 }
 
-type Result struct {
-	Error    string `name:"error"`
+type LocalProc struct {
 	Pid      int64  `name:"pid"`
 	Result   string `name:"result"`
 	ExitCode int64  `name:"exit_code"`
+}
+
+type Result struct {
+	Error string    `name:"error"`
+	Proc  LocalProc `name:"proc"`
 }
