+++ functions/sshconf/types.go
@@ -22,6 +22,18 @@ var (
 	AgentIdentifier = "crashd_ssh_agent"
 )
 
+// Args represent input arguments passed to starlark function.
+// Args can also be used as output arguments to built-in function.
+//
+// The argument map follows:
+//   - username - username
+//   - port string - SSH port
+//   - private_key_path string - SSH private key path
+//   - jump_user string - jump host username
+//   - jump_host string - jump host name
+//   - max_retries [] string - maximum retires for SSH
+//   - conn_timeout bool - timeout for connection
+//
 type Args struct {
 	Username       string `name:"username"`
 	Port           string `name:"port" optional:"true"`
@@ -32,8 +44,8 @@ type Args struct {
 	ConnTimeout    int64  `name:"conn_timeout" optional:"true"`
 }
 
+// Config is a configuration returned by the command function
 type Config struct {
-	Error          string `name:"error"`
 	Username       string `name:"username"`
 	Port           string `name:"port"`
 	PrivateKeyPath string `name:"private_key_path"`
@@ -41,4 +53,9 @@ type Config struct {
 	JumpHost       string `name:"jump_host"`
 	MaxRetries     int64  `name:"max_retries"`
 	ConnTimeout    int64  `name:"conn_timeout"`
+}
+
+type Result struct {
+	Error string `name:"error"`
+	Conf  Config `name:"conf"`
 }
